subscriber/requirements.txt
# Runtime dependencies
pymongo[srv]==3.10.1
paho-mqtt==1.5.0
requests==2.21.0